+++ CHANGES.md
@@ -19,6 +19,7 @@ master (in development)
   - Fix offset of column numbers in ESLint [GH-640]
   - Properly parse indentation errors from Python 2.7 [GH-635]
   - Don’t choke if `default-directory` does not exist [GH-625]
+  - Fix error parsing for Puppet 4
 
 0.23 (Apr 6, 2015)
 ==================